requirements.txt
Flask==3.1.2
Jinja2~=3.1.6
gunicorn==24.1.1
Markdown==3.10.1
mysql-connector-python==9.5.0
numpy==2.4.1
pytz==2025.2
